requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
Flask==2.0.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.3
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scipy==1.6.3
seaborn==0.11.1
six==1.16.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.1
